feat: optimize theme-token components
